+++ packages/jss/tests/jss-tests.ts
@@ -1,12 +1,33 @@
 // API docs at http://cssinjs.org/js-api
 
-import {create as createJSS, createGenerateId, SheetsRegistry, default as sharedInstance} from 'jss'
+import {
+  create as createJSS,
+  createGenerateId,
+  JssStyle,
+  SheetsRegistry,
+  default as sharedInstance
+} from 'jss'
+import {NextChannel} from 'indefinite-observable'
 
 const jss = createJSS().setup({createGenerateId})
 jss.use({}, {}) // $ExpectType JSS
 
 const styleSheet = jss.createStyleSheet<string>(
   {
+    ruleWithMockObservable: {
+      subscribe: (observer: {next: NextChannel<JssStyle | string | null | undefined>}) => {
+        const next = typeof observer === 'function' ? observer : observer.next
+        next({background: 'blue', display: 'flex'})
+
+        // These tests are ported over from https://github.com/DefinitelyTyped/DefinitelyTyped/blob/de655960b603d6b47f7030674f084780c76e045f/types/jss/jss-tests.ts
+        // where there were $ExpectError cases; however, those don't seem to be
+        // enforced in JSS's current testing harness.
+
+        return {
+          unsubscribe() {}
+        }
+      }
+    },
     rule: {
       color: (data: {color: string}) => data.color,
 
@@ -29,6 +50,7 @@ const styleSheet = jss.createStyleSheet<string>(
 const attachedStyleSheet = styleSheet.attach()
 
 attachedStyleSheet.classes.container // $ExpectType string
+attachedStyleSheet.classes.ruleWithMockObservable // $ExpectType string
 
 const rule = attachedStyleSheet.addRule('dynamicRule', {color: 'indigo'})
 attachedStyleSheet.classes.dynamicRule // $ExpectType string
@@ -57,6 +79,13 @@ attachedStyleSheet.addRules({
 const styleSheet2 = sharedInstance.createStyleSheet({
   container: {
     background: '#000099'
+  },
+  ruleWithMockObservable: {
+    subscribe() {
+      return {
+        unsubscribe() {}
+      }
+    }
   }
 })